Add a way to create custom retry policies.
A RetryStrategy will dictate what decision will be made based on the result of the sent request.
